requirements.txt
absl-py==1.4.0
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==7.1.2
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.35
distlib==0.3.6
filelock==3.12.0
Flask==1.1.2
Flask-Cors==4.0.0
fonttools==4.39.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.2
google-auth-oauthlib==0.4.6
grpcio==1.51.3
gunicorn==20.0.4
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==1.1.1
matplotlib==3.7.1
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.1
pandas==1.5.3
Pillow==9.4.0
platformdirs==3.5.1
pluggy==1.0.0
protobuf==4.22.1
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pyproject-api==1.5.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
pyzbar==0.1.9
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tesseract==0.1.3
thop==0.1.1.post2209072238
tomli==2.0.1
torch==1.9.1
torchaudio==0.9.1
torchvision==0.10.1
tox==4.5.2
tqdm==4.65.0
typing-extensions==4.5.0
urllib3==1.26.15
virtualenv==20.23.0
Werkzeug==1.0.1
zipp==3.15.0